requirements.txt
absl-py==0.8.1
asn1crypto==0.24.0
astor==0.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cryptography==2.1.4
filelock==3.0.12
Flask==1.1.2
gast==0.2.2
gdown==3.12.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.17.3
opensimplex==0.3
opt-einsum==3.1.0
Pillow==7.2.0
protobuf==3.10.0
pycrypto==2.6.1
pygobject==3.26.1
PySocks==1.7.1
python-apt==1.6.4
pyxdg==0.25
requests==2.24.0
scipy==1.5.2
SecretStorage==2.3.1
six==1.11.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==0.16.0
wrapt==1.11.2
